fix: use `error_description` instead of `error.message`
The commit replaces the usage of `error.message` with `error_description`
in `Vault.php` and the `GetSecretTest.php` and `SetSecretTest.php` tests
to match the Azure Key Vault API response. It fixes the handling of
errors when retrieving and setting secrets.

The commit also updates the
error message in the tests to use the example error response from the Azure documentation: https://learn.microsoft.com/en-us/azure/active-directory/develop/v2-oauth2-auth-code-flow#error-response-1